# docs/swagger.yaml
definitions:
...
  lifecycle.approveCCRequest:
    properties:
      cc_name:
        type: string
      cc_sequence:
        type: integer
      cc_version:
        type: string
      channel_name:
        type: string
      package_ID:
        type: string
    type: object
  lifecycle.approvedChaincodeResponse:
    properties:
      endorsement_plugin:
        type: string
      init_required:
        type: boolean
      package_ID:
        type: string
      sequence:
        type: integer
      validation_plugin:
        type: string
      version:
        type: string
    type: object
  lifecycle.ccApprovalRequest:
    properties:
      cc_name:
        type: string
      cc_sequence:
        type: integer
      cc_version:
        type: string
      channel_name:
        type: string
    type: object
  lifecycle.ccApprovals:
    properties:
      approvals:
        additionalProperties:
          type: boolean
        type: object
    type: object
  lifecycle.commitCCRequest:
    properties:
      cc_name:
        type: string
      cc_sequence:
        type: integer
      cc_version:
        type: string
      channel_name:
        type: string
    type: object
  lifecycle.committedChaincodeResponse:
    properties:
      chaincode:
        type: string
      channel:
        type: string
      endorsement_plugin:
        type: string
      sequence:
        type: string
      validation_plugin:
        type: string
      version:
        type: string
    type: object
...
  lifecycle.queryRequest:
    properties:
      cc_name:
        type: string
      channel_name:
        type: string
    type: object
...
  repository.cloneRequest:
    properties:
      directory:
        type: string
      url:
        type: string
    type: object
...
info:
  contact:
    name: arogya.Gurkha
    url: https://github.com/arogyaGurkha
  description: API to run fabric binaries
  title: Hyperledger Fabric Gurkhaman API
  version: "0.1"
paths:
...
  /fabric/lifecycle/approve:
    get:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode queryapproved` is executed through `exec.Command()`
        to query approved chaincode definitions.'
      parameters:
      - description: cc name and the channel it was approved in
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.queryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/lifecycle.approvedChaincodeResponse'
      summary: Query an approved chaincode definition on a channel.
      tags:
      - lifecycle
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode approveformyorg` is executed through
        `exec.Command()` to approve a chaincode definition.'
      parameters:
      - description: channel name (mychannel), cc name (basic), cc version (1.0),
          cc sequence (1), package ID (run [GET] /fabric/lifecycle/install)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.approveCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Approve the cc definition for the current org.
      tags:
      - lifecycle
  /fabric/lifecycle/commit:
    get:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode querycommited` is executed through `exec.Command()`
        to query committed chaincode definitions.'
      parameters:
      - description: cc name and the channel it was committed in
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.queryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/lifecycle.committedChaincodeResponse'
      summary: Query the committed chaincode definitions by channel on a peer.
      tags:
      - lifecycle
    post:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode commit` is executed through `exec.Command()`
        to commit chaincode definition on a channel.'
      parameters:
      - description: channel name (mychannel), cc name (basic), cc version (1.0),
          cc sequence (1)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.commitCCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Commit the chaincode definition on the channel.
      tags:
      - lifecycle
  /fabric/lifecycle/commit/organizations:
    get:
      consumes:
      - application/json
      description: '`peer lifecycle chaincode checkcommitreadiness` is executed through
        `exec.Command()` to check commit readiness.'
      parameters:
      - description: channel name (mychannel), cc name (basic), cc version (1.0),
          cc sequence (1)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/lifecycle.ccApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/lifecycle.ccApprovals'
      summary: Check whether a chaincode definition is ready to be committed on a
        channel. Shows which organizations have approved the cc definition.
      tags:
      - lifecycle
...
  /fabric/repository/clone:
    get:
      consumes:
      - application/json
      description: Clone a repository.
      parameters:
      - description: url (https://github.com/arogyaGurkha/GurkhaContracts.git), directory
          (GurkhaContracts or nil)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/repository.cloneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successful operation
      summary: Clone a repository.
      tags:
      - repository
...
swagger: "2.0"